Crear entidad Repuesto de Maquinas, seteado min y max tipofalla, fix calcular tiempo de reparacion

+++ backend/src/main/java/com/example/metalTest/mantenimientoCorrectivo/domain/MantenimientoCorrectivo.java
@@ -52,7 +52,7 @@ public class MantenimientoCorrectivo {
     private int nrocorrectivo;
 
     @Column
-    private int tiempoReparacion;
+    private long tiempoReparacion;
 
     @JoinColumn
     @OneToOne
